+++ src/main/java/neat/XOR3I.java
@@ -18,45 +18,45 @@ public class XOR3I implements FitnessFunction{
 	}
 
 	@Override
-	public float calculateFitness(Genome gnm) {
-		List<Float> inList = new ArrayList<Float>();
-		inList.add(1f);
-		inList.add(1f);
-		inList.add(1f);
-		float d1 = gnm.calculate(inList).get(0);
-		inList.set(0, 0f);
-		inList.set(1, 0f);
-		inList.set(2, 0f);
-		float d2 = gnm.calculate(inList).get(0);
-		inList.set(0, 0f);
-		inList.set(1, 0f);
-		inList.set(2, 1f);
-		float d3 = 1f - gnm.calculate(inList).get(0);
-		inList.set(0, 0f);
-		inList.set(1, 1f);
-		inList.set(2, 0f);
-		float d4 = 1f - gnm.calculate(inList).get(0);
-		inList.set(0, 0f);
-		inList.set(1, 1f);
-		inList.set(2, 1f);
-		float d5 = 1f - gnm.calculate(inList).get(0);
-		inList.set(0, 1f);
-		inList.set(1, 0f);
-		inList.set(2, 0f);
-		float d6 = 1f - gnm.calculate(inList).get(0);
-		inList.set(0, 1f);
-		inList.set(1, 0f);
-		inList.set(2, 1f);
-		float d7 = 1f - gnm.calculate(inList).get(0);
-		inList.set(0, 1f);
-		inList.set(1, 1f);
-		inList.set(2, 0f);
-		float d8 = 1f - gnm.calculate(inList).get(0);
-		float d = d1 + d2 + d3 + d4 + d5 + d6 + d7 + d8;
-		return (8f - d) * (8f - d);
+	public double calculateFitness(Genome gnm) {
+		List<Double> inList = new ArrayList<Double>();
+		inList.add(1d);
+		inList.add(1d);
+		inList.add(1d);
+		double d1 = gnm.calculate(inList).get(0);
+		inList.set(0, 0d);
+		inList.set(1, 0d);
+		inList.set(2, 0d);
+		double d2 = gnm.calculate(inList).get(0);
+		inList.set(0, 0d);
+		inList.set(1, 0d);
+		inList.set(2, 1d);
+		double d3 = 1d - gnm.calculate(inList).get(0);
+		inList.set(0, 0d);
+		inList.set(1, 1d);
+		inList.set(2, 0d);
+		double d4 = 1d - gnm.calculate(inList).get(0);
+		inList.set(0, 0d);
+		inList.set(1, 1d);
+		inList.set(2, 1d);
+		double d5 = 1d - gnm.calculate(inList).get(0);
+		inList.set(0, 1d);
+		inList.set(1, 0d);
+		inList.set(2, 0d);
+		double d6 = 1d - gnm.calculate(inList).get(0);
+		inList.set(0, 1d);
+		inList.set(1, 0d);
+		inList.set(2, 1d);
+		double d7 = 1d - gnm.calculate(inList).get(0);
+		inList.set(0, 1d);
+		inList.set(1, 1d);
+		inList.set(2, 0d);
+		double d8 = 1d - gnm.calculate(inList).get(0);
+		double d = d1 + d2 + d3 + d4 + d5 + d6 + d7 + d8;
+		return (8d - d) * (8d - d);
 	}
 	@Override
-	public float getThreshold() {
-		return 63f;
+	public double getThreshold() {
+		return 63d;
 	}
 }
